network: avoid DHCP
DHCP slow thing down a lot (=~15s in a 9 nodes env). We this patch, the
nodes network configuration is static.

The static IP addresses are recorded in the domain XMLs, this way, we
can do quickly list the IP already in use, even if the domain are not
running.

Also:
- Centos-7: enable NetworkManager to be consistent with the RHEL-7
  images
- Adjust the default configuration of the images to prevent any DHCP
  auto-configuration

Others:
- do not call `systemctl mask cloud-init`, this was breaking cloud-init
  at last in one environment.
- fix a little performance issue with the SSH ping