recycle previous MAC
Time to time, a VM is deleted, but its entry in the DHCP leases remains.
As a result, the DHCP won't give back the same IP until the lease as
expired.
We now reuse previously the same MAC for the new instance as soon as the
IP and the hostname are the same.