add images.json and refresh the image lists
The commit introduces a new file called `images.json` that will later
replace the Caddy webserver. We will modify virt-lightning fetch in
order to retrieve this `images.json` and avoid the complex system of HTTP
redirection that we have today.

This commits also remove a series of outdated images (`404`).

Also, import `refresh.py` is a help script used to maintain the different files.